Refactor app into ES modules
