+++ index.html
@@ -41,7 +41,7 @@
     <button onclick="addRule()">Add New Rule</button>
   </section>
 
-  <script src="app.js?v=2"></script>
+  <script type="module" src="src/main.js"></script>
 </body>
 
 </html>
