+++ package.json
@@ -1,0 +1,4 @@
+{
+  "name": "timesheetapp",
+  "type": "module"
+}
