+++ test/parseSummary.test.js
@@ -1,13 +1,13 @@
-const test = require('node:test');
-const assert = require('node:assert');
+import test from 'node:test';
+import assert from 'node:assert';
 
-// Polyfill minimal localStorage before requiring app.js
+// Polyfill minimal localStorage before importing modules
 global.localStorage = {
   getItem: () => null,
   setItem: () => {}
 };
 
-const { parseSummary } = require('../app.js');
+const { parseSummary } = await import('../src/parser.js');
 
 const sampleDay = 'Mon';
